Looping composite activities
Transitions now support a 'reinitialize' property, which will cause
CompositeActivities to reinitialize their child workflow when executed.
Added a 'Loop' transition to the default CompositeActivity, which
connects started back to itself with the new 'reinitialize' flag set,
and the enablingProperty field set to 'RepeatWhen'.
Enhanced the reading of such activity properties to enable the use of
DataHelpers, so that the Loop transition (or any enablingProperty value)
can be dependent on outcome fields or property values.
Cleaned up a lot of arcane run() and runNext() logic.

Fixes #1